extract rpc clients from proposer into its own crate under crates/proof
