+++ Cargo.toml
@@ -26,6 +26,7 @@ members = [
   "crates/proof/tee/client",
   "crates/proof/tee/server",
   "crates/proof/contracts",
+  "crates/proof/rpc",
   "crates/proof/proposer",
   "crates/utilities/*",
   "crates/builder/*",
@@ -215,6 +216,7 @@ base-enclave = { path = "crates/proof/tee/core" }
 base-enclave-client = { path = "crates/proof/tee/client" }
 base-enclave-server = { path = "crates/proof/tee/server" }
 base-proof-contracts = { path = "crates/proof/contracts" }
+base-proof-rpc = { path = "crates/proof/rpc" }
 base-proposer = { path = "crates/proof/proposer" }
 
 # Utilities
