+++ crates/proof/proposer/src/constants.rs
@@ -29,9 +29,6 @@ pub const ECDSA_V_OFFSET: u8 = 27;
 /// Default poll interval for checking new blocks.
 pub const DEFAULT_POLL_INTERVAL: Duration = Duration::from_secs(12);
 
-/// Default LRU cache size for RPC responses.
-pub const DEFAULT_CACHE_SIZE: usize = 1000;
-
 /// Sentinel value for the parent game index when creating the first game from
 /// the anchor state registry (i.e., no parent game exists).
 /// This is `uint32.max` per the `DisputeGameFactory` contract.
@@ -53,19 +50,6 @@ pub const PROOF_TYPE_TEE: u8 = 0;
 /// further to ensure the proposer can always find and resume from its most recent
 /// game after a restart.
 pub const MAX_GAME_RECOVERY_LOOKBACK: u64 = 5000;
-
-// ============================================================================
-// Retry Configuration Constants
-// ============================================================================
-
-/// Default maximum number of retry attempts for RPC operations.
-pub const DEFAULT_RPC_MAX_RETRIES: u32 = 5;
-
-/// Default initial delay for exponential backoff.
-pub const DEFAULT_RETRY_INITIAL_DELAY: Duration = Duration::from_millis(100);
-
-/// Default maximum delay between retry attempts.
-pub const DEFAULT_RETRY_MAX_DELAY: Duration = Duration::from_secs(10);
 
 // ============================================================================
 // Gas Estimation Constants
