+++ crates/proof/rpc/Cargo.toml
@@ -1,0 +1,51 @@
+[package]
+name = "base-proof-rpc"
+description = "Shared RPC client abstractions for OP Stack L1, L2, and rollup node interactions"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+# Alloy
+alloy-transport-http.workspace = true
+alloy-rpc-types-eth.workspace = true
+alloy-rpc-client.workspace = true
+alloy-primitives.workspace = true
+alloy-transport.workspace = true
+alloy-provider.workspace = true
+alloy-network.workspace = true
+alloy-eips.workspace = true
+
+# Base Alloy
+base-alloy-rpc-types.workspace = true
+base-alloy-network.workspace = true
+
+# Enclave
+base-enclave.workspace = true
+
+# Async
+async-trait.workspace = true
+
+# Tracing
+tracing.workspace = true
+
+# Serialization
+serde_json.workspace = true
+serde.workspace = true
+
+# Error handling
+thiserror.workspace = true
+
+# Misc
+moka = { workspace = true, features = ["future"] }
+backon.workspace = true
+url.workspace = true
+
+[dev-dependencies]
+tokio = { workspace = true, features = ["test-util"] }
